client: relocated some components and updated some css
