+++ client/src/components/SubmitButton/SubmitButton.css.js
@@ -1,13 +1,13 @@
 import { makeStyles } from "@material-ui/core";
 
-const useStyles = makeStyles({
+const useStyles = makeStyles((theme) => ({
   button: {
     borderRadius: 20,
     color: "#FFFFFF",
-    backgroundColor: "#43DDC1",
+    backgroundColor: `${theme.palette.secondary.main}`,
     width: "15ch",
     padding: 10,
   },
-});
+}));
 
 export default useStyles;
