+++ client/src/pages/CodeUploadDialog/CodeUploadDialog.css.js
@@ -1,17 +1,17 @@
 import { makeStyles } from "@material-ui/core/styles";
 
-const useStyles = makeStyles({
+const useStyles = makeStyles((theme) => ({
   form: {
     "& label.Mui-focused": {
-      color: "purple",
+      color: theme.palette.primary.dark,
     },
 
     "& .MuiOutlinedInput-root": {
       "&.Mui-focused fieldset": {
-        borderColor: "purple",
+        borderColor: theme.palette.primary.dark,
       },
     },
   },
-});
+}));
 
 export default useStyles;
