+++ client/src/pages/Login.jsx
@@ -3,12 +3,12 @@ import { Link } from "react-router-dom";
 import axios from "axios";
 import {
   makeStyles,
-  Button,
   TextField,
   FormHelperText,
 } from "@material-ui/core";
 import LoginSignupContainer from "components/LoginSignupContainer";
 import GridTemplateContainer from "components/GridTemplateContainer";
+import SubmitButton from 'components/SubmitButton';
 import { store } from "utils/storage";
 
 // TODO Figure out where to move useStyles to avoid duplicate code
@@ -29,11 +29,6 @@ const useStyles = makeStyles((theme) => ({
   button: {
     marginTop: 30,
     marginBottom: 60,
-    borderRadius: 20,
-    color: "#FFFFFF",
-    backgroundColor: `${theme.palette.secondary.main}`,
-    width: "15ch",
-    padding: 10,
   },
   textfield: {
     "& .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-notchedOutline": {
@@ -132,9 +127,9 @@ const Login = () => {
             </FormHelperText>
           )}
 
-          <Button type="submit" variant="contained" className={classes.button}>
+          <SubmitButton className={classes.button}>
             Login
-          </Button>
+          </SubmitButton>
 
           <div>
             <strong>
