+++ client/src/utils/storage.js
@@ -1,3 +1,7 @@
-export const store = (token) => (localStorage.token = token);
+export const store = (token) => {
+  localStorage.token = token;
+};
 
-export const remove = (token) => localStorage.removeItem(token);
+export const remove = (token) => {
+  localStorage.removeItem(token);
+};
